# docs/swagger.yaml
basePath: /
definitions:
  entity.Config:
    properties:
      createdBy:
        type: string
      createdDate:
        type: string
      id:
        type: string
      isSystem:
        description: 1=是 0=否
        type: integer
      key:
        type: string
      name:
        type: string
      remark:
        type: string
      updatedBy:
        type: string
      updatedDate:
        type: string
      value:
        type: string
    type: object
  entity.Dept:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.Dept'
        type: array
      createdBy:
        type: string
      createdDate:
        type: string
      email:
        maxLength: 100
        type: string
      id:
        type: string
      leader:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      orderNum:
        minimum: 0
        type: integer
      parentId:
        type: string
      phone:
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      updatedBy:
        type: string
      updatedDate:
        type: string
    required:
    - name
    type: object
  entity.Job:
    properties:
      beanName:
        maxLength: 100
        minLength: 2
        type: string
      createdTime:
        type: string
      cronExpression:
        maxLength: 100
        type: string
      id:
        type: string
      methodName:
        maxLength: 100
        minLength: 2
        type: string
      params:
        maxLength: 500
        type: string
      remark:
        maxLength: 500
        type: string
      status:
        description: 0:正常 1:暂停
        enum:
        - 0
        - 1
        type: integer
    required:
    - beanName
    - cronExpression
    - methodName
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      restCode:
        type: string
    type: object
  v1.LoginRequest:
    properties:
      loginName:
        description: 必填校验
        type: string
      password:
        type: string
    required:
    - loginName
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: simonchen453@gmail.com
    name: API Support
  description: 基于 Clean Architecture 的企业级后台管理系统 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Pro Golang API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - auth
...
  /api/v1/generator/batchGenCode:
    get:
      parameters:
      - description: 表名列表(逗号分隔)
        in: query
        name: tables
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 批量生成代码
      tags:
      - generator
  /api/v1/generator/list:
    get:
      parameters:
      - description: 表名
        in: query
        name: tableName
        type: string
      - description: 页码
        in: query
        name: pageNo
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取数据表列表
      tags:
      - generator
  /api/v1/job:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务信息
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/entity.Job'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增定时任务
      tags:
      - job
    put:
      consumes:
      - application/json
      parameters:
      - description: 任务信息
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/entity.Job'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新定时任务
      tags:
      - job
  /api/v1/job/{id}:
    delete:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除定时任务
      tags:
      - job
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取定时任务详情
      tags:
      - job
...
  /api/v1/job/log/{id}:
    get:
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取任务日志详情
      tags:
      - job
...
  /api/v1/system/config:
    post:
      consumes:
      - application/json
      parameters:
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/entity.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增参数配置
      tags:
      - system/config
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/entity.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新参数配置
      tags:
      - system/config
  /api/v1/system/config/{id}:
    delete:
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除参数配置
      tags:
      - system/config
    get:
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取参数配置详情
      tags:
      - system/config
  /api/v1/system/config/key/{key}:
    get:
      parameters:
      - description: 配置Key
        in: path
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据Key获取配置
      tags:
      - system/config
...
  /api/v1/system/dept:
    post:
      consumes:
      - application/json
      parameters:
      - description: 部门信息
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/entity.Dept'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增部门
      tags:
      - system/dept
    put:
      consumes:
      - application/json
      parameters:
      - description: 部门信息
        in: body
        name: dept
        required: true
        schema:
          $ref: '#/definitions/entity.Dept'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新部门
      tags:
      - system/dept
  /api/v1/system/dept/{id}:
    delete:
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除部门
      tags:
      - system/dept
    get:
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取部门详情
      tags:
      - system/dept
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"